# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Navigant Education Consultants API",
    "version": "1.0.0",
    "description": "API for tracking community empowerment tasks focused on SDG 4 (Quality Education) and SDG 5 (Gender Equality)."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/tasks": {
      "get": {
        "summary": "Get all Navigant tasks",
        "responses": {
          "200": {
            "description": "A successful list of operations"
          }
        }
      },
      "post": {
        "summary": "Add a new community training task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "category": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task successfully created"
          }
        }
      }
    },
    "/tasks/{id}": {
      "put": {
        "summary": "Update an existing task status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "completed": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task successfully updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a completed task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Task successfully deleted"
          }
        }
      }
    }
  }
}